Add redux logger to mobile web in dev environment
