+++ web/static/mobile_survey/js/store/configureStore.dev.js
@@ -1,0 +1,6 @@
+import prodStore from './configureStore.prod'
+import createLogger from 'redux-logger'
+
+export default function configureStore(preloadedState) {
+  return prodStore(preloadedState, [createLogger()])
+}
